move level2 widgets into a "widgets" directory and namespace
