+++ Makefile
@@ -18,12 +18,12 @@ textbox.cpp                            \
 level2/input.cpp                       \
 level2/application.cpp                 \
 level2/drawing_context_proxy.cpp       \
-level2/label.cpp                       \
-level2/button.cpp                      \
-level2/checkbox.cpp                    \
-level2/radio_button.cpp                \
-level2/scrolling_list.cpp              \
-level2/simple_plot.cpp                 \
+level2/widgets/label.cpp               \
+level2/widgets/button.cpp              \
+level2/widgets/checkbox.cpp            \
+level2/widgets/radio_button.cpp        \
+level2/widgets/scrolling_list.cpp      \
+level2/widgets/simple_plot.cpp         \
 drivers/display_stm3210e-eval.cpp      \
 drivers/event_stm3210e-eval.cpp        \
 drivers/display_mp3v2.cpp              \
