+++ level2/widgets/checkbox.h
@@ -31,7 +31,7 @@
 
 #ifdef MXGUI_LEVEL_2
 
-namespace mxgui {
+namespace mxgui::widgets {
 
 /**
  * CheckBox Button.
