+++ level2/widgets/simple_plot.cpp
@@ -33,7 +33,7 @@
 
 using namespace std;
 
-namespace mxgui {
+namespace mxgui::widgets {
 
 //
 // class SimplePlot
